bazci: move the logic of posting github issues to bazci
This code change moves the logic of filtering tests JSON
output files and posting github issues to bazci.

Release note: None
