+++ pkg/BUILD.bazel
@@ -112,6 +112,8 @@ ALL_TESTS = [
     "//pkg/cloud/userfile/filetable/filetabletest:filetabletest_test",
     "//pkg/cloud/userfile:userfile_test",
     "//pkg/clusterversion:clusterversion_test",
+    "//pkg/cmd/bazci/githubpost:githubpost_test",
+    "//pkg/cmd/bazci/testfilter:testfilter_test",
     "//pkg/cmd/cmpconn:cmpconn_test",
     "//pkg/cmd/cockroach-oss:cockroach-oss_disallowed_imports_test",
     "//pkg/cmd/dev:dev_lib_disallowed_imports_test",
@@ -119,7 +121,6 @@ ALL_TESTS = [
     "//pkg/cmd/docgen/extract:extract_test",
     "//pkg/cmd/docs-issue-generation:docs-issue-generation_test",
     "//pkg/cmd/generate-bazel-extra:generate-bazel-extra_test",
-    "//pkg/cmd/github-post:github-post_test",
     "//pkg/cmd/github-pull-request-make:github-pull-request-make_test",
     "//pkg/cmd/internal/issues:issues_test",
     "//pkg/cmd/label-merged-pr:label-merged-pr_test",
@@ -135,7 +136,6 @@ ALL_TESTS = [
     "//pkg/cmd/roachtest/tests:tests_test",
     "//pkg/cmd/roachtest:roachtest_test",
     "//pkg/cmd/teamcity-trigger:teamcity-trigger_test",
-    "//pkg/cmd/testfilter:testfilter_test",
     "//pkg/col/coldata:coldata_disallowed_imports_test",
     "//pkg/col/coldata:coldata_test",
     "//pkg/col/coldataext:coldataext_test",
@@ -862,6 +862,10 @@ GO_TARGETS = [
     "//pkg/clusterversion:clusterversion_test",
     "//pkg/cmd/allocsim:allocsim",
     "//pkg/cmd/allocsim:allocsim_lib",
+    "//pkg/cmd/bazci/githubpost:githubpost",
+    "//pkg/cmd/bazci/githubpost:githubpost_test",
+    "//pkg/cmd/bazci/testfilter:testfilter",
+    "//pkg/cmd/bazci/testfilter:testfilter_test",
     "//pkg/cmd/bazci:bazci",
     "//pkg/cmd/bazci:bazci_lib",
     "//pkg/cmd/cmdutil:cmdutil",
@@ -917,7 +921,6 @@ GO_TARGETS = [
     "//pkg/cmd/geoviz:geoviz_lib",
     "//pkg/cmd/github-post:github-post",
     "//pkg/cmd/github-post:github-post_lib",
-    "//pkg/cmd/github-post:github-post_test",
     "//pkg/cmd/github-pull-request-make:github-pull-request-make",
     "//pkg/cmd/github-pull-request-make:github-pull-request-make_lib",
     "//pkg/cmd/github-pull-request-make:github-pull-request-make_test",
@@ -993,7 +996,6 @@ GO_TARGETS = [
     "//pkg/cmd/teamcity-trigger:teamcity-trigger_test",
     "//pkg/cmd/testfilter:testfilter",
     "//pkg/cmd/testfilter:testfilter_lib",
-    "//pkg/cmd/testfilter:testfilter_test",
     "//pkg/cmd/uptodate:uptodate",
     "//pkg/cmd/uptodate:uptodate_lib",
     "//pkg/cmd/urlcheck/lib/urlcheck:urlcheck",
@@ -2300,6 +2302,8 @@ GET_X_DATA_TARGETS = [
     "//pkg/clusterversion:get_x_data",
     "//pkg/cmd/allocsim:get_x_data",
     "//pkg/cmd/bazci:get_x_data",
+    "//pkg/cmd/bazci/githubpost:get_x_data",
+    "//pkg/cmd/bazci/testfilter:get_x_data",
     "//pkg/cmd/cmdutil:get_x_data",
     "//pkg/cmd/cmp-protocol:get_x_data",
     "//pkg/cmd/cmp-protocol/pgconnect:get_x_data",
