+++ pkg/cmd/bazci/BUILD.bazel
@@ -12,6 +12,8 @@ go_library(
     deps = [
         "//pkg/build/bazel/bes",
         "//pkg/build/util",
+        "//pkg/cmd/bazci/githubpost",
+        "//pkg/cmd/bazci/testfilter",
         "@com_github_alessio_shellescape//:shellescape",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_gogo_protobuf//proto",
