+++ pkg/cmd/bazci/githubpost/BUILD.bazel
@@ -1,0 +1,32 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "githubpost",
+    srcs = ["githubpost.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/bazci/githubpost",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/cmd/internal/issues",
+        "//pkg/internal/codeowners",
+        "//pkg/internal/team",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
+
+go_test(
+    name = "githubpost_test",
+    srcs = ["githubpost_test.go"],
+    args = ["-test.timeout=295s"],
+    data = glob(["testdata/**"]),
+    embed = [":githubpost"],
+    tags = ["broken_in_bazel"],
+    deps = [
+        "//pkg/cmd/internal/issues",
+        "//pkg/testutils",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+    ],
+)
+
+get_x_data(name = "get_x_data")
