+++ pkg/cmd/bazci/githubpost/testdata/implicit-pkg.json
