+++ pkg/cmd/bazci/githubpost/testdata/stress-failure.json
