+++ pkg/cmd/bazci/githubpost/testdata/stress-fatal.json
