+++ pkg/cmd/bazci/githubpost/testdata/stress-init-panic.json
