+++ pkg/cmd/bazci/githubpost/testdata/stress-panic.json
