+++ pkg/cmd/bazci/testfilter/BUILD.bazel
@@ -1,0 +1,25 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "testfilter",
+    srcs = ["testfilter.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/bazci/testfilter",
+    visibility = ["//visibility:public"],
+    deps = ["@org_golang_x_sync//errgroup"],
+)
+
+go_test(
+    name = "testfilter_test",
+    srcs = ["testfilter_test.go"],
+    args = ["-test.timeout=295s"],
+    data = glob(["testdata/**"]),
+    embed = [":testfilter"],
+    deps = [
+        "//pkg/testutils",
+        "//pkg/util/leaktest",
+        "@com_github_cockroachdb_datadriven//:datadriven",
+    ],
+)
+
+get_x_data(name = "get_x_data")
