+++ pkg/cmd/bazci/testfilter/testdata/strip.txt
